requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6
openai==1.3.0
pymongo==4.6.2
motor==3.3.2
beanie==1.25.0
python-jose[cryptography]==3.3.0
certifi==2024.2.2 